提交用户，组，类型服务类代码

+++ starter/src/main/java/starter/rest/Systems.java
@@ -88,6 +88,7 @@ public class Systems {
                            @RequestParam(defaultValue = "10") int limit,
                            @RequestParam(defaultValue = "[]") SortBuilder[] sort) {
         try {
+            checkAuthorize();
             XContentBuilder result = userService.all(query, start, limit, sort);
             return result.string();
         } catch (IOException e) {
@@ -95,19 +96,10 @@ public class Systems {
         }
     }
 
-//    @RequestMapping(value = "users", method = RequestMethod.GET)
-//    public String allUsers() {
-//        try {
-//            XContentBuilder result = userService.all();
-//            return result.string();
-//        } catch (IOException e) {
-//            throw new uContentException(e, HttpStatus.INTERNAL_SERVER_ERROR);
-//        }
-//    }
-
     @RequestMapping(value = "users", method = RequestMethod.POST)
     public String createUser(@RequestBody Json body) {
         try {
+            checkAuthorize();
             XContentBuilder result = userService.create(body);
             return result.string();
         } catch (IOException e) {
@@ -118,6 +110,7 @@ public class Systems {
     @RequestMapping(value = "users/{id}", method = RequestMethod.GET)
     public String getUser(@PathVariable String id) {
         try {
+            checkAuthorize();
             XContentBuilder result = userService.get(id).toXContentBuilder();
             return result.string();
         } catch (IOException e) {
@@ -129,6 +122,7 @@ public class Systems {
     @RequestMapping(value = "users/{id}", method = RequestMethod.PATCH)
     public String updateUser(@PathVariable String id,@RequestBody Json body) {
         try {
+            checkAuthorize();
             XContentBuilder result = userService.update(id, body);
             return result.string();
         } catch (IOException e) {
@@ -140,6 +134,7 @@ public class Systems {
     @RequestMapping(value = "users/{id}", method = RequestMethod.DELETE)
     public String deleteUser(@PathVariable String id) {
         try {
+            checkAuthorize();
             XContentBuilder result = userService.delete(id);
             return result.string();
         } catch (IOException e) {
@@ -150,6 +145,7 @@ public class Systems {
     @RequestMapping(value = "users/{id}/groups", method = RequestMethod.GET)
     public String getUserGroups(@PathVariable String id) {
         try {
+            checkAuthorize();
             XContentBuilder result = userService.getGroups(id);
             return result.string();
         } catch (IOException e) {
@@ -160,22 +156,13 @@ public class Systems {
 
     /****************************** groups ******************************/
 
-//    @RequestMapping(value = "groups", method = RequestMethod.GET)
-//    public String allGroups() {
-//        try {
-//            XContentBuilder result = groupService.all();
-//            return result.string();
-//        } catch (IOException e) {
-//            throw new uContentException(e, HttpStatus.INTERNAL_SERVER_ERROR);
-//        }
-//    }
-
     @RequestMapping(value = "groups", method = {RequestMethod.GET})
     public String allGroups(@RequestParam(defaultValue = "") String query,
                             @RequestParam(defaultValue = "0") int start,
                             @RequestParam(defaultValue = "10") int limit,
                             @RequestParam(defaultValue = "[]") SortBuilder[] sort) {
         try {
+            checkAuthorize();
             XContentBuilder result = groupService.all(query, start, limit, sort);
             return result.string();
         } catch (IOException e) {
@@ -186,6 +173,7 @@ public class Systems {
     @RequestMapping(value = "groups", method = RequestMethod.POST)
     public String createGroup(@RequestBody Json body) {
         try {
+            checkAuthorize();
             XContentBuilder result = groupService.create(body);
             return result.string();
         } catch (IOException e) {
@@ -197,6 +185,7 @@ public class Systems {
     public String refGroupUsers(@PathVariable String id,
                                  @RequestBody Json userIds) {
         try {
+            checkAuthorize();
             XContentBuilder result = groupService.refUsers(id, userIds);
             return result.string();
         } catch (IOException e) {
@@ -207,6 +196,7 @@ public class Systems {
     @RequestMapping(value = "groups/{id}", method = RequestMethod.GET)
     public String getGroup(@PathVariable String id) {
         try {
+            checkAuthorize();
             XContentBuilder result = groupService.get(id).toXContentBuilder();
             return result.string();
         } catch (IOException e) {
@@ -217,6 +207,7 @@ public class Systems {
     @RequestMapping(value = "groups/{id}/users", method = RequestMethod.GET)
     public String getGroupUsers(@PathVariable String id) {
         try {
+            checkAuthorize();
             XContentBuilder result = groupService.getUsers(id);
             return result.string();
         } catch (IOException e) {
@@ -227,6 +218,7 @@ public class Systems {
     @RequestMapping(value = "groups/{id}", method = RequestMethod.PATCH)
     public String updateGroup(@PathVariable String id,@RequestBody Json body) {
         try {
+            checkAuthorize();
             XContentBuilder result = groupService.update(id, body);
             return result.string();
         } catch (IOException e) {
@@ -237,6 +229,7 @@ public class Systems {
     @RequestMapping(value = "groups/{id}", method = RequestMethod.DELETE)
     public String deleteGroup(@PathVariable String id) {
         try {
+            checkAuthorize();
             XContentBuilder result = groupService.delete(id);
             return result.string();
         } catch (IOException e) {
@@ -258,5 +251,12 @@ public class Systems {
     @RequestMapping(value = "_reIndex", method = RequestMethod.POST)
     public void reindex() {
         reIndexService.reIndex();
+    }
+
+
+    private void checkAuthorize(){
+        if(!groupService.checkUserInAdminGroup()){
+            throw new uContentException("Forbidden", HttpStatus.FORBIDDEN);
+        }
     }
 }
